Add solver_info returned from a solver operator apply.
This adds a struct containing information returned from the apply of
solver operators.  This struct currently relays the status (whether
the apply returned due to convergence or divergence) and the number of
iterations completed.

The BiCGSTAB test was changed to test for convergence in a set number
of iterations for the given matrices.  This required adding a seed
option for the simple `set_random` vector operation.